[N-04] Explicitly use uint256 in type declarations

+++ contracts/MintableToken.sol
@@ -117,7 +117,7 @@ contract MintableToken is
     {
         uint256 currentAllowance = allowance(_msgSender(), spender);
         require(
-            type(uint).max - currentAllowance >= addedValue,
+            type(uint256).max - currentAllowance >= addedValue,
             ErrorCoded.ERR_ARITHMETIC_OVERFLOW_ALLOWANCE
         );
         unchecked {
